YYC-110: split tui/state.rs — extract domain types
tui/state.rs grew to ~1545 lines mixing AppState, ChatMessage and its
impl block, SessionState, orchestration tracking, picker state, and
the entire test suite. Carve out the domain types into focused
sibling modules so each surface is independently navigable.

* tui/chat_message.rs — MessageSegment, ChatMessage struct, and every
  method on it (new, append_text, append_reasoning, push_tool_start,
  finish_tool, render_version, ...).
* tui/orchestration.rs — SubAgentTile, ToolLogRow, TickerCell, TreeNode,
  OrchestrationPhase + impl, OrchestrationEvent, OrchestrationState +
  Default impl. Currently demo-shaped; the runtime they need is
  YYC-68.
* tui/picker_state.rs — SessionState, SessionStatus, ProviderPickerEntry.
* tui/state/tests.rs — full test suite, declared via `#[cfg(test)]
  mod tests;`. State.rs converted to state/ directory module.

state/mod.rs re-exports the moved types via `pub use` so existing
`tui::state::ChatMessage` import paths across the TUI keep working.

Public API unchanged. cargo test --all-features --lib passes
(252 tests).

Note on size: the issue's <600 target on state/mod.rs would require
splitting impl AppState across multiple files (inherent impl
allowed in sibling modules). That's a deeper refactor than the
structural split scoped here; state/mod.rs lands at 826 lines.
Tracking the AppState method split as a follow-up.
