+++ src/tui/state/mod.rs
@@ -24,6 +24,15 @@ use ratatui::style::Color;
 use crate::hooks::audit::{AuditBuffer, AuditKind};
 use crate::memory::SessionSummary;
 
+// Re-export the types pulled into sibling submodules in YYC-110 so the
+// legacy `tui::state::*` import paths used across the TUI still resolve.
+pub use super::chat_message::{ChatMessage, MessageSegment};
+pub use super::orchestration::{
+    OrchestrationEvent, OrchestrationPhase, OrchestrationState, SubAgentTile, TickerCell,
+    ToolLogRow, TreeNode,
+};
+pub use super::picker_state::{ProviderPickerEntry, SessionState, SessionStatus};
+
 use super::theme::{Palette, Theme};
 use super::views::{DiffKind, DiffLine, View};
 
@@ -108,367 +117,6 @@ pub enum ToolStatus {
     InProgress,
     /// `bool` is the success flag — true → ✓, false → ✗.
     Done(bool),
-}
-
-/// One slice of an agent's response timeline, in arrival order. Segments
-/// preserve the natural interleaving of reasoning, tool calls, and text so
-/// the renderer can show the agent's actual flow (think → tool → think →
-/// answer) instead of bunching all reasoning above all tool calls (YYC-71).
-#[derive(Clone, Debug)]
-pub enum MessageSegment {
-    Reasoning(String),
-    Text(String),
-    /// One tool invocation rendered as a structured card (YYC-74).
-    /// `params_summary` is the one-line projection from the agent's
-    /// `summarize_tool_args` (e.g. path for file ops, query for
-    /// search). `output_preview` is a truncated tail of the tool
-    /// result. `elapsed_ms` is wall-clock dispatch time for the
-    /// timing note. All optional — older streams that don't populate
-    /// them still render a minimal card.
-    ToolCall {
-        name: String,
-        status: ToolStatus,
-        params_summary: Option<String>,
-        output_preview: Option<String>,
-        /// One-line metadata derived from tool result (e.g. "847 lines",
-        /// "5 matches", "+12 -3"). Renders as a dimmed sub-header in
-        /// the YYC-74 card.
-        result_meta: Option<String>,
-        /// Lines elided beyond the preview (YYC-78). When > 0 the card
-        /// renders a "… N more lines elided" footer.
-        elided_lines: usize,
-        elapsed_ms: Option<u64>,
-    },
-}
-
-impl MessageSegment {
-    /// Stable kind tag used by the renderer to detect transitions between
-    /// segment types and insert a blank-line separator (YYC-91).
-    pub fn kind_label(&self) -> &'static str {
-        match self {
-            MessageSegment::Reasoning(_) => "reasoning",
-            MessageSegment::Text(_) => "text",
-            MessageSegment::ToolCall { .. } => "tool",
-        }
-    }
-}
-
-#[derive(Clone, Debug, Default)]
-pub struct ChatMessage {
-    pub role: ChatRole,
-    /// Flattened text content (kept for hydration from `Message::Assistant`
-    /// where the wire format only has aggregate content + reasoning, no
-    /// per-event timeline). Used as a fallback by the renderer when
-    /// `segments` is empty.
-    pub content: String,
-    /// Aggregate reasoning trace for hydrated messages. Renderer falls back
-    /// to this only when `segments` is empty.
-    pub reasoning: String,
-    /// Ordered timeline of reasoning fragments, tool calls, and text chunks
-    /// as they arrived. Populated for live streamed agent messages; empty
-    /// for user/system messages and hydrated history.
-    pub segments: Vec<MessageSegment>,
-    pub(crate) render_version: u64,
-}
-
-impl ChatMessage {
-    pub fn new(role: ChatRole, content: impl Into<String>) -> Self {
-        Self {
-            role,
-            content: content.into(),
-            reasoning: String::new(),
-            segments: Vec::new(),
-            render_version: 0,
-        }
-    }
-
-    pub fn render_version(&self) -> u64 {
-        self.render_version
-    }
-
-    fn bump_render_version(&mut self) {
-        self.render_version = self.render_version.wrapping_add(1);
-    }
-
-    pub fn set_content(&mut self, content: impl Into<String>) {
-        self.content = content.into();
-        self.bump_render_version();
-    }
-
-    /// True if neither the live segment timeline nor the hydrated content
-    /// field has any text. Used by the renderer to decide whether to show
-    /// the streaming "Thinking…" / "Answering…" placeholder.
-    pub fn has_text(&self) -> bool {
-        if !self.content.is_empty() {
-            return true;
-        }
-        self.segments
-            .iter()
-            .any(|s| matches!(s, MessageSegment::Text(t) if !t.is_empty()))
-    }
-
-    /// True if any reasoning has been recorded — either streamed into
-    /// segments or hydrated into the legacy `reasoning` field.
-    pub fn has_reasoning(&self) -> bool {
-        if !self.reasoning.is_empty() {
-            return true;
-        }
-        self.segments
-            .iter()
-            .any(|s| matches!(s, MessageSegment::Reasoning(r) if !r.is_empty()))
-    }
-
-    /// Append a text chunk to the segment timeline, coalescing with the
-    /// trailing segment if it's also text. Leading whitespace on a fresh
-    /// text segment (start of message, or after a tool/reasoning break)
-    /// is stripped so models that emit `\n\n` preambles don't render
-    /// gaps before the visible body.
-    pub fn append_text(&mut self, chunk: &str) {
-        match self.segments.last_mut() {
-            Some(MessageSegment::Text(t)) => {
-                t.push_str(chunk);
-                self.bump_render_version();
-            }
-            _ => {
-                let trimmed = chunk.trim_start_matches(['\n', '\r']);
-                if trimmed.is_empty() {
-                    return;
-                }
-                self.segments
-                    .push(MessageSegment::Text(trimmed.to_string()));
-                self.bump_render_version();
-            }
-        }
-    }
-
-    /// Append a reasoning chunk to the segment timeline, coalescing with the
-    /// trailing segment if it's also reasoning. New tool calls or text break
-    /// the run, so subsequent reasoning starts a fresh segment — that's the
-    /// whole point of YYC-71.
-    ///
-    /// Whitespace-only chunks are dropped so empty `<think></think>` blocks
-    /// from local providers (YYC-103) don't create empty reasoning segments
-    /// that the renderer would later show as a stub THINKING header
-    /// between adjacent text deltas.
-    pub fn append_reasoning(&mut self, chunk: &str) {
-        if chunk.trim().is_empty() {
-            return;
-        }
-        match self.segments.last_mut() {
-            Some(MessageSegment::Reasoning(r)) => r.push_str(chunk),
-            _ => self
-                .segments
-                .push(MessageSegment::Reasoning(chunk.to_string())),
-        }
-        self.bump_render_version();
-    }
-
-    pub fn push_tool_start(&mut self, name: impl Into<String>) {
-        self.push_tool_start_with(name, None);
-    }
-
-    /// Push a tool-call segment with the params summary the YYC-74
-    /// card needs.
-    pub fn push_tool_start_with(
-        &mut self,
-        name: impl Into<String>,
-        params_summary: Option<String>,
-    ) {
-        self.segments.push(MessageSegment::ToolCall {
-            name: name.into(),
-            status: ToolStatus::InProgress,
-            params_summary,
-            output_preview: None,
-            result_meta: None,
-            elided_lines: 0,
-            elapsed_ms: None,
-        });
-        self.bump_render_version();
-    }
-
-    /// Mark the most recent in-progress ToolCall with this name as done.
-    /// Walks segments in reverse so concurrent dispatch (YYC-34) still
-    /// pairs each end with its own start as the matching tail.
-    pub fn finish_tool(&mut self, name: &str, ok: bool) {
-        self.finish_tool_with(name, ok, None, None, 0, None);
-    }
-
-    /// Same as `finish_tool` but also stamps the result preview, meta
-    /// summary, elided count, and timing for the YYC-74 card.
-    pub fn finish_tool_with(
-        &mut self,
-        name: &str,
-        ok: bool,
-        output_preview: Option<String>,
-        result_meta: Option<String>,
-        elided_lines: usize,
-        elapsed_ms: Option<u64>,
-    ) {
-        for seg in self.segments.iter_mut().rev() {
-            if let MessageSegment::ToolCall {
-                name: n,
-                status,
-                output_preview: op,
-                result_meta: rm,
-                elided_lines: el,
-                elapsed_ms: em,
-                ..
-            } = seg
-                && n == name && matches!(status, ToolStatus::InProgress) {
-                    *status = ToolStatus::Done(ok);
-                    *op = output_preview;
-                    *rm = result_meta;
-                    *el = elided_lines;
-                    *em = elapsed_ms;
-                    self.bump_render_version();
-                    return;
-                }
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SessionState {
-    pub id: String,
-    pub label: String,
-    pub message_count: usize,
-    pub created_at: i64,
-    pub last_active: i64,
-    pub parent_session_id: Option<String>,
-    pub lineage_label: Option<String>,
-    pub preview: Option<String>,
-    pub status: SessionStatus,
-    pub is_active: bool,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum SessionStatus {
-    Live,
-    Saved,
-}
-
-impl SessionStatus {
-    pub fn label(self) -> &'static str {
-        match self {
-            Self::Live => "live",
-            Self::Saved => "saved",
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SubAgentTile {
-    pub name: String,
-    pub role: String,
-    pub state: String,
-    pub color: Color,
-    pub log: Vec<String>,
-    pub cpu: Vec<u16>,
-}
-
-#[derive(Clone, Debug)]
-pub struct ToolLogRow {
-    pub time: String,
-    pub actor: String,
-    pub msg: String,
-}
-
-/// One row in the provider picker overlay (YYC-97). `None` name = the
-/// legacy unnamed `[provider]` block; `Some(name)` = a `[providers.<name>]`
-/// profile.
-#[derive(Clone, Debug)]
-pub struct ProviderPickerEntry {
-    pub name: Option<String>,
-    pub model: String,
-    pub base_url: String,
-}
-
-#[derive(Clone, Debug)]
-pub struct TickerCell {
-    pub sub: String,
-    pub msg: String,
-    pub color: Color,
-}
-
-#[derive(Clone, Debug)]
-pub struct TreeNode {
-    pub depth: u8,
-    pub label: String,
-    pub state: String,
-    pub color: Color,
-    pub active: bool,
-}
-
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq)]
-pub enum OrchestrationPhase {
-    #[default]
-    Idle,
-    Thinking,
-    ToolRunning,
-    Paused,
-    Complete,
-    Error,
-}
-
-impl OrchestrationPhase {
-    fn label(self) -> &'static str {
-        match self {
-            Self::Idle => "idle",
-            Self::Thinking => "thinking",
-            Self::ToolRunning => "tool",
-            Self::Paused => "paused",
-            Self::Complete => "done",
-            Self::Error => "error",
-        }
-    }
-
-    fn color(self) -> Color {
-        match self {
-            Self::Idle => Palette::MUTED,
-            Self::Thinking => Palette::YELLOW,
-            Self::ToolRunning => Palette::BLUE,
-            Self::Paused => Palette::RED,
-            Self::Complete => Palette::GREEN,
-            Self::Error => Palette::RED,
-        }
-    }
-
-    fn symbol(self) -> &'static str {
-        match self {
-            Self::Idle => "○",
-            Self::Thinking => "●",
-            Self::ToolRunning => "●",
-            Self::Paused => "⏸",
-            Self::Complete => "✓",
-            Self::Error => "✗",
-        }
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct OrchestrationEvent {
-    pub actor: String,
-    pub msg: String,
-    pub color: Color,
-}
-
-#[derive(Clone, Debug)]
-pub struct OrchestrationState {
-    pub phase: OrchestrationPhase,
-    pub active_task: String,
-    pub current_tool: Option<String>,
-    pub recent_events: Vec<OrchestrationEvent>,
-}
-
-impl Default for OrchestrationState {
-    fn default() -> Self {
-        Self {
-            phase: OrchestrationPhase::Idle,
-            active_task: "Awaiting user input".into(),
-            current_tool: None,
-            recent_events: Vec::new(),
-        }
-    }
 }
 
 pub struct AppState {
@@ -1188,392 +836,4 @@ fn short_text(text: &str, max: usize) -> String {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::memory::SessionSummary;
-
-    #[test]
-    fn orchestration_state_tracks_prompt_and_tool_flow() {
-        let mut app = AppState::new("test-model".into(), 128_000);
-        app.note_prompt_submitted("list files in the current directory");
-        assert_eq!(app.orchestration.phase, OrchestrationPhase::Thinking);
-        assert!(app.orchestration.active_task.contains("list files"));
-
-        app.note_tool_start("bash");
-        assert_eq!(app.orchestration.phase, OrchestrationPhase::ToolRunning);
-        assert_eq!(app.orchestration.current_tool.as_deref(), Some("bash"));
-
-        let tree = app.tree_nodes();
-        assert_eq!(tree.len(), 2);
-        assert!(tree[1].label.contains("bash"));
-
-        app.note_tool_end("bash", true);
-        assert_eq!(app.orchestration.phase, OrchestrationPhase::Thinking);
-        assert!(app.orchestration.current_tool.is_none());
-
-        app.note_done();
-        assert_eq!(app.orchestration.phase, OrchestrationPhase::Complete);
-        assert_eq!(app.ticker_cells().last().unwrap().msg, "completed turn");
-    }
-
-    #[test]
-    fn subagent_tiles_expose_single_real_orchestrator() {
-        let mut app = AppState::new("test-model".into(), 128_000);
-        app.note_prompt_submitted("check auth middleware");
-        let tiles = app.subagent_tiles();
-        assert_eq!(tiles.len(), 1);
-        assert_eq!(tiles[0].name, "main");
-        assert_eq!(tiles[0].role, "orchestrator");
-        assert!(tiles[0].log[0].contains("check auth middleware"));
-    }
-
-    #[test]
-    fn hydrate_sessions_retains_lineage_and_activity_fields() {
-        let mut app = AppState::new("test-model".into(), 128_000);
-        let parent_id = "parent-12345678".to_string();
-        let child_id = "child-87654321".to_string();
-        app.hydrate_sessions(
-            &[SessionSummary {
-                id: child_id.clone(),
-                created_at: 10,
-                last_active: 20,
-                message_count: 3,
-                parent_session_id: Some(parent_id.clone()),
-                lineage_label: Some("branched from auth cleanup".into()),
-                provider_profile: None,
-                preview: None,
-            }],
-            &child_id,
-        );
-
-        assert_eq!(app.sessions.len(), 1);
-        let session = &app.sessions[0];
-        assert_eq!(session.id, child_id);
-        assert_eq!(
-            session.parent_session_id.as_deref(),
-            Some(parent_id.as_str())
-        );
-        assert_eq!(
-            session.lineage_label.as_deref(),
-            Some("branched from auth cleanup")
-        );
-        assert_eq!(session.status, SessionStatus::Live);
-        assert!(session.is_active);
-        assert_eq!(session.message_count, 3);
-    }
-
-    #[test]
-    fn segments_interleave_reasoning_tool_text_in_arrival_order() {
-        // Simulates the exact YYC-71 sequence: think → tool → think → answer.
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        m.append_reasoning("checking the file");
-        m.push_tool_start("read_file");
-        m.finish_tool("read_file", true);
-        m.append_reasoning("now writing");
-        m.push_tool_start("write_file");
-        m.finish_tool("write_file", true);
-        m.append_text("Done!");
-
-        let kinds: Vec<&str> = m
-            .segments
-            .iter()
-            .map(|s| match s {
-                MessageSegment::Reasoning(_) => "reasoning",
-                MessageSegment::Text(_) => "text",
-                MessageSegment::ToolCall { .. } => "tool",
-            })
-            .collect();
-        assert_eq!(
-            kinds,
-            vec!["reasoning", "tool", "reasoning", "tool", "text"]
-        );
-    }
-
-    #[test]
-    fn append_reasoning_coalesces_until_broken_by_other_segment() {
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        m.append_reasoning("first ");
-        m.append_reasoning("chunk");
-        m.push_tool_start("bash");
-        m.append_reasoning("after tool");
-        // Three segments: reasoning, tool, reasoning — second reasoning is
-        // its own segment because the tool call broke the run.
-        assert_eq!(m.segments.len(), 3);
-        match &m.segments[0] {
-            MessageSegment::Reasoning(r) => assert_eq!(r, "first chunk"),
-            other => panic!("expected reasoning, got {other:?}"),
-        }
-        match &m.segments[2] {
-            MessageSegment::Reasoning(r) => assert_eq!(r, "after tool"),
-            other => panic!("expected reasoning, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn chat_message_render_version_bumps_on_mutation() {
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        let initial = m.render_version();
-
-        m.append_text("hello");
-        assert!(m.render_version() > initial);
-        let after_text = m.render_version();
-
-        m.append_reasoning("thinking");
-        assert!(m.render_version() > after_text);
-        let after_reasoning = m.render_version();
-
-        m.push_tool_start("bash");
-        assert!(m.render_version() > after_reasoning);
-        let after_tool_start = m.render_version();
-
-        m.finish_tool("bash", true);
-        assert!(m.render_version() > after_tool_start);
-    }
-
-    #[test]
-    fn model_status_omits_prefix_when_no_provider_label() {
-        let mut app = AppState::new("deepseek/v4".into(), 128_000);
-        app.prompt_tokens_last = 18_402;
-        let status = app.model_status();
-        assert_eq!(status, "deepseek/v4 · 18,402 / 128,000");
-        assert!(!status.contains(" · deepseek/v4"));
-    }
-
-    #[test]
-    fn model_status_prefixes_active_provider_label() {
-        let mut app = AppState::new("deepseek/v4".into(), 128_000);
-        app.provider_label = Some("local".into());
-        app.prompt_tokens_last = 18_402;
-        let status = app.model_status();
-        assert_eq!(status, "local · deepseek/v4 · 18,402 / 128,000");
-    }
-
-    #[test]
-    fn chat_message_new_starts_at_zero_render_version() {
-        let m = ChatMessage::new(ChatRole::User, "hello");
-        assert_eq!(m.render_version(), 0);
-    }
-
-    #[test]
-    fn refresh_prompt_mode_picks_command_when_input_starts_with_slash() {
-        let mut app = AppState::new("test".into(), 100);
-        app.input = "/help".into();
-        app.refresh_prompt_mode();
-        assert_eq!(app.prompt_mode, PromptMode::Command);
-    }
-
-    #[test]
-    fn refresh_prompt_mode_returns_to_insert_when_slash_cleared() {
-        let mut app = AppState::new("test".into(), 100);
-        app.input = "/help".into();
-        app.refresh_prompt_mode();
-        app.input.clear();
-        app.refresh_prompt_mode();
-        assert_eq!(app.prompt_mode, PromptMode::Insert);
-    }
-
-    #[test]
-    fn refresh_prompt_mode_uses_busy_when_thinking() {
-        let mut app = AppState::new("test".into(), 100);
-        app.thinking = true;
-        app.refresh_prompt_mode();
-        assert_eq!(app.prompt_mode, PromptMode::Busy);
-        assert_eq!(app.mode_label(), "BUSY");
-    }
-
-    #[test]
-    fn refresh_prompt_mode_busy_overrides_command_prefix() {
-        // While the agent is mid-turn the badge should still read BUSY
-        // even if the user typed `/` in the prompt — the slash menu can
-        // be shown but the mode pill reflects the agent state.
-        let mut app = AppState::new("test".into(), 100);
-        app.thinking = true;
-        app.input = "/queue".into();
-        app.refresh_prompt_mode();
-        assert_eq!(app.prompt_mode, PromptMode::Busy);
-    }
-
-    #[test]
-    fn prompt_hints_default_keybinds_match_ascii_labels() {
-        // Default Keybinds should produce ASCII-safe labels (Ctrl+T,
-        // Ctrl+K) so prompt-row chips render in any terminal font.
-        let app = AppState::new("test".into(), 100);
-        let hints = app.prompt_hints();
-        let pairs: Vec<(String, String)> = hints.to_vec();
-        assert!(
-            pairs.contains(&("Ctrl+T".into(), "tools".into())),
-            "expected Ctrl+T tools in {pairs:?}"
-        );
-        assert!(
-            pairs.contains(&("Ctrl+K".into(), "sessions".into())),
-            "expected Ctrl+K sessions in {pairs:?}"
-        );
-    }
-
-    #[test]
-    fn prompt_hints_reflect_overridden_keybind() {
-        use super::super::keybinds::{KeyBinding, Keybinds};
-        use crossterm::event::{KeyCode, KeyModifiers};
-        let mut kb = Keybinds::defaults();
-        kb.toggle_tools = KeyBinding {
-            code: KeyCode::F(2),
-            mods: KeyModifiers::NONE,
-        };
-        let app = AppState::new("test".into(), 100).with_keybinds(kb);
-        let pairs: Vec<(String, String)> = app.prompt_hints().to_vec();
-        assert!(
-            pairs.contains(&("F2".into(), "tools".into())),
-            "expected F2 tools in {pairs:?}"
-        );
-        assert!(
-            !pairs.iter().any(|(k, _)| k == "Ctrl+T"),
-            "stale Ctrl+T label leaked into {pairs:?}"
-        );
-    }
-
-    #[test]
-    fn prompt_hints_returns_borrowed_slice_no_alloc() {
-        // Two consecutive calls in the same mode must return slices to
-        // the same cached storage — proves we're not allocating per call.
-        let app = AppState::new("test".into(), 100);
-        let first = app.prompt_hints().as_ptr();
-        let second = app.prompt_hints().as_ptr();
-        assert_eq!(first, second, "prompt_hints reallocated between calls");
-    }
-
-    #[test]
-    fn format_thousands_groups_at_three_digit_boundaries() {
-        assert_eq!(format_thousands(0), "0");
-        assert_eq!(format_thousands(42), "42");
-        assert_eq!(format_thousands(999), "999");
-        assert_eq!(format_thousands(1_000), "1,000");
-        assert_eq!(format_thousands(18_402), "18,402");
-        assert_eq!(format_thousands(1_234_567), "1,234,567");
-    }
-
-    #[test]
-    fn context_ratio_reflects_latest_prompt_size() {
-        let mut app = AppState::new("test".into(), 100_000);
-        assert_eq!(app.context_ratio(), 0.0);
-        app.prompt_tokens_last = 50_000;
-        assert!((app.context_ratio() - 0.5).abs() < 1e-6);
-        app.prompt_tokens_last = 95_000;
-        assert!(app.context_ratio() > 0.9);
-    }
-
-    #[test]
-    fn estimated_cost_returns_none_without_pricing() {
-        let mut app = AppState::new("test".into(), 100_000);
-        app.prompt_tokens_total = 1_000;
-        app.completion_tokens_total = 500;
-        assert!(app.estimated_cost().is_none());
-    }
-
-    #[test]
-    fn estimated_cost_multiplies_tokens_by_per_token_pricing() {
-        let mut app = AppState::new("test".into(), 100_000);
-        app.prompt_tokens_total = 1_000;
-        app.completion_tokens_total = 500;
-        app.pricing = Some(crate::provider::catalog::Pricing {
-            input_per_token: 0.000_001,
-            output_per_token: 0.000_002,
-        });
-        let cost = app.estimated_cost().unwrap();
-        // 1000*0.000001 + 500*0.000002 = 0.001 + 0.001 = 0.002
-        assert!((cost - 0.002).abs() < 1e-9, "got {cost}");
-    }
-
-    #[test]
-    fn lifetime_tokens_sums_prompt_and_completion_totals() {
-        let mut app = AppState::new("test".into(), 100_000);
-        app.prompt_tokens_total = 1_000;
-        app.completion_tokens_total = 500;
-        assert_eq!(app.lifetime_tokens(), 1_500);
-    }
-
-    #[test]
-    fn queue_starts_empty_and_pushes_in_fifo_order() {
-        let mut app = AppState::new("test".into(), 100);
-        assert!(app.queue.is_empty());
-        app.queue.push_back("first".into());
-        app.queue.push_back("second".into());
-        assert_eq!(app.queue.pop_front().as_deref(), Some("first"));
-        assert_eq!(app.queue.pop_front().as_deref(), Some("second"));
-        assert!(app.queue.is_empty());
-    }
-
-    #[test]
-    fn queue_pop_back_removes_most_recent_only() {
-        let mut app = AppState::new("test".into(), 100);
-        app.queue.push_back("a".into());
-        app.queue.push_back("b".into());
-        app.queue.push_back("c".into());
-        app.queue.pop_back();
-        let remaining: Vec<&str> = app.queue.iter().map(String::as_str).collect();
-        assert_eq!(remaining, vec!["a", "b"]);
-    }
-
-    #[test]
-    fn finish_tool_pairs_with_most_recent_in_progress_call_of_same_name() {
-        // Parallel dispatch (YYC-34): two write_file calls in flight; the
-        // first to finish should pair with the most recent matching start
-        // that's still in-progress.
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        m.push_tool_start("write_file");
-        m.push_tool_start("write_file");
-        m.finish_tool("write_file", true);
-
-        let statuses: Vec<ToolStatus> = m
-            .segments
-            .iter()
-            .filter_map(|s| match s {
-                MessageSegment::ToolCall { status, .. } => Some(*status),
-                _ => None,
-            })
-            .collect();
-        // Most-recent in-progress finishes first; the older one stays open.
-        assert_eq!(
-            statuses,
-            vec![ToolStatus::InProgress, ToolStatus::Done(true)]
-        );
-    }
-
-    #[test]
-    fn push_tool_start_with_carries_params_summary() {
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        m.push_tool_start_with("read_file", Some("src/foo.rs".into()));
-        match &m.segments[0] {
-            MessageSegment::ToolCall { params_summary, .. } => {
-                assert_eq!(params_summary.as_deref(), Some("src/foo.rs"))
-            }
-            other => panic!("expected ToolCall, got {other:?}"),
-        }
-    }
-
-    #[test]
-    fn finish_tool_with_stamps_preview_and_timing() {
-        let mut m = ChatMessage::new(ChatRole::Agent, "");
-        m.push_tool_start("read_file");
-        m.finish_tool_with(
-            "read_file",
-            true,
-            Some("hello\nworld".into()),
-            Some("2 lines".into()),
-            0,
-            Some(345),
-        );
-        match &m.segments[0] {
-            MessageSegment::ToolCall {
-                output_preview,
-                elapsed_ms,
-                status,
-                ..
-            } => {
-                assert!(matches!(status, ToolStatus::Done(true)));
-                assert_eq!(output_preview.as_deref(), Some("hello\nworld"));
-                assert_eq!(*elapsed_ms, Some(345));
-            }
-            other => panic!("expected ToolCall, got {other:?}"),
-        }
-    }
-}
+mod tests;
